transform ?tronky into ?treat
